Issue #21: Validation code moved and duplicate code eliminated.

+++ src/Services/RouteActivator.php
@@ -9,8 +9,10 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Log;
 use TromsFylkestrafikk\Netex\Models\ActiveJourney;
 use TromsFylkestrafikk\Netex\Models\ActiveCall;
+use TromsFylkestrafikk\Netex\Services\RouteBase;
+use TromsFylkestrafikk\Netex\Services\RouteValidator;
 
-class RouteActivator
+class RouteActivator extends RouteBase
 {
     /**
      * @var string
@@ -21,6 +23,11 @@ class RouteActivator
      * @var string
      */
     protected $toDate;
+
+    /**
+     * @var string[]
+     */
+    protected $activationDates = [];
 
     /**
      * @var \TromsFylkestrafikk\Netex\Services\DbBulkInsert
@@ -139,26 +146,28 @@ class RouteActivator
         $prevJourneyCount = 0;
         $prevCallCount = 0;
         while ($date <= $toDate) {
-            // Reset internal overview of seen IDs
-            $this->callIds = [];
-            $this->aJourneyIds = [];
             $dateStr = $date->format('Y-m-d');
-            $rawJourneys = $this->getRawJourneys($dateStr);
-            $this->activateJourneys($dateStr, $rawJourneys);
-            $this->dayCount++;
-            // Interim flush to assert the full day is complete, and get the
-            // correct written count.
-            $this->journeyDumper->flush();
-            $this->callDumper->flush();
+            if (in_array($dateStr, $this->activationDates)) {
+                // Reset internal overview of seen IDs
+                $this->callIds = [];
+                $this->aJourneyIds = [];
+                $rawJourneys = parent::getRawJourneys($dateStr);
+                $this->activateJourneys($dateStr, $rawJourneys);
+                $this->dayCount++;
+                // Interim flush to assert the full day is complete, and get the
+                // correct written count.
+                $this->journeyDumper->flush();
+                $this->callDumper->flush();
+                Log::debug(sprintf(
+                    "NeTEx: %s: %d journeys and %d calls",
+                    $dateStr,
+                    $this->journeyDumper->getRecordsWritten() - $prevJourneyCount,
+                    $this->callDumper->getRecordsWritten() - $prevCallCount,
+                ));
+                $prevJourneyCount = $this->journeyDumper->getRecordsWritten();
+                $prevCallCount = $this->callDumper->getRecordsWritten();
+            }
             $this->invoke($this->dayCallback, $dateStr);
-            Log::debug(sprintf(
-                "NeTEx: %s: %d journeys and %d calls",
-                $dateStr,
-                $this->journeyDumper->getRecordsWritten() - $prevJourneyCount,
-                $this->callDumper->getRecordsWritten() - $prevCallCount,
-            ));
-            $prevJourneyCount = $this->journeyDumper->getRecordsWritten();
-            $prevCallCount = $this->callDumper->getRecordsWritten();
             $date->addDay();
         }
         // Write last prepared records;
@@ -178,24 +187,60 @@ class RouteActivator
      *
      * @return $this
      */
-    public function deactivate()
+    public function deactivate($forced = false)
     {
         $date = new Carbon($this->fromDate);
         $toDate = new Carbon($this->toDate);
 
         $this->dayCount = 0;
 
-        Log::info(sprintf("NeTEx: De-activating route data between %s and %s", $this->fromDate, $this->toDate));
+        Log::info(sprintf("NeTEx: Deactivating route data between %s and %s", $this->fromDate, $this->toDate));
         while ($date <= $toDate) {
-            DB::table('netex_active_calls', 'call')
-                ->join('netex_active_journeys as journey', 'call.active_journey_id', '=', 'journey.id')
-                ->whereDate('journey.date', $date)->delete();
-            DB::table('netex_active_journeys')->whereDate('date', $date)->delete();
-            $this->dayCount++;
-            $this->invoke($this->dayCallback, $date->format('Y-m-d'));
+            $dateStr = $date->format('Y-m-d');
+            if ($forced || in_array($dateStr, $this->activationDates)) {
+                DB::table('netex_active_calls', 'call')
+                    ->join('netex_active_journeys as journey', 'call.active_journey_id', '=', 'journey.id')
+                    ->whereDate('journey.date', $date)->delete();
+                DB::table('netex_active_journeys')->whereDate('date', $date)->delete();
+                Log::debug(sprintf("NeTEx: %s: Deactivated.", $dateStr));
+                $this->dayCount++;
+            }
+            $this->invoke($this->dayCallback, $dateStr);
             $date->addDay();
         }
-        Log::info("NeTEx: De-activating complete");
+        Log::info(sprintf("NeTEx: Deactivation %s",
+            ($this->dayCount > 0) ? 'complete.' : 'skipped.',
+        ));
+        return $this;
+    }
+
+    /**
+     * Validate route set for the given period against already activated data.
+     *
+     * @return $this
+     */
+    public function validate()
+    {
+        $date = new Carbon($this->fromDate);
+        $toDate = new Carbon($this->toDate);
+
+        $validator = new RouteValidator();
+        Log::info(sprintf("NeTEx: Validating route data between %s and %s", $this->fromDate, $this->toDate));
+
+        while ($date <= $toDate) {
+            $dateStr = $date->format('Y-m-d');
+            $mismatch = $validator->validateJourneys($dateStr);
+            if ($mismatch) {
+                $this->activationDates[] = $dateStr;
+            } else {
+                Log::debug(sprintf("NeTEx: Content for %s is matching existing data.", $dateStr));
+            }
+            $this->invoke($this->dayCallback, $dateStr);
+            $date->addDay();
+        }
+        Log::info(sprintf("NeTEx: Validation complete. Activation required for %d date(s).",
+            count($this->activationDates),
+        ));
         return $this;
     }
 
@@ -240,35 +285,12 @@ class RouteActivator
         ];
     }
 
-    public function makeCallId(array $callRecord, $journeyRecord)
-    {
-        return (is_array($journeyRecord)
-                ? $this->makeJourneyId($journeyRecord)
-                : $journeyRecord)
-            . ':'
-            . $callRecord['order'];
-    }
-
-    /**
-     * @param mixed[] $journeyRecord
-     *
-     * @return string
-     */
-    public function makeJourneyId(array $journeyRecord)
-    {
-        return implode(':', [
-            $journeyRecord['date'],
-            $journeyRecord['line_private_code'],
-            $journeyRecord['private_code'],
-        ]);
-    }
-
     protected function activateJourneys($date, Collection $rawJourneys)
     {
         foreach ($rawJourneys as $rawJourney) {
             $jRec = array_intersect_key((array) $rawJourney, $this->journeyFillable);
             $jRec['date'] = $date;
-            $jId = $this->makeJourneyId($jRec);
+            $jId = parent::makeJourneyId($jRec);
             if (!empty($this->aJourneyIds[$jId])) {
                 Log::error(sprintf(
                     'NeTEx: Duplicate active journey ID detected: %s. Journey ID: %s (%s)',
@@ -289,12 +311,12 @@ class RouteActivator
 
     protected function activateJourneyCalls(array &$jRec)
     {
-        $rawCalls = $this->getRawCalls($jRec['vehicle_journey_id']);
+        $rawCalls = parent::getRawCalls($jRec['vehicle_journey_id']);
         $callStamp = new Carbon("{$jRec['date']} 04:00:00");
         $prevDestDisplay = $jRec['name'];
         foreach ($rawCalls as $rawCall) {
-            $callStamp = $this->expandCallTime($rawCall, 'arrival_time', $callStamp);
-            $callStamp = $this->expandCallTime($rawCall, 'departure_time', $callStamp);
+            $callStamp = parent::expandCallTime($rawCall, 'arrival_time', $callStamp);
+            $callStamp = parent::expandCallTime($rawCall, 'departure_time', $callStamp);
             if ($rawCall->destination_display) {
                 $prevDestDisplay = $rawCall->destination_display;
             } else {
@@ -312,32 +334,6 @@ class RouteActivator
     }
 
     /**
-     * Expand a time only ('HH:mm:ss') time format to full date timestamp.
-     *
-     * For calls that passes midnight, we need to keep track of the date and
-     * update it when needed. This updated Carbon timestamp is returned.
-     *
-     * @param \stdClass &$rawCall
-     * @param string $property
-     * @param \Illuminate\Support\Carbon $prevCallStamp
-     *
-     * @return \Illuminate\Support\Carbon
-     */
-    protected function expandCallTime(&$rawCall, $property, $prevCallStamp)
-    {
-        if (!$rawCall->$property) {
-            return $prevCallStamp;
-        }
-        $dateStr = $prevCallStamp->format('Y-m-d');
-        $callStamp = new Carbon("$dateStr {$rawCall->$property}");
-        if ($callStamp < $prevCallStamp) {
-            $callStamp->addDay();
-        }
-        $rawCall->$property = $callStamp->format('Y-m-d H:i:s');
-        return $callStamp;
-    }
-
-    /**
      * Activate (persistent store) a single raw call.
      *
      * @param mixed[] $jRec
@@ -345,7 +341,7 @@ class RouteActivator
      */
     protected function activateCall(array $jRec, array $rawCall)
     {
-        $callId = $this->makeCallId($rawCall, $jRec['id']);
+        $callId = parent::makeCallId($rawCall, $jRec['id']);
         if (!empty($this->callIds[$callId])) {
             Log::error(sprintf(
                 "NeTEx activation: Duplicate active call detected: %s on journey ID %s. Call time: %s (%s)",
@@ -361,77 +357,11 @@ class RouteActivator
         $this->callDumper->addRecord(array_merge(
             array_intersect_key($rawCall, $this->callFillable),
             [
-                'id' => $this->makeCallId($rawCall, $jRec['id']),
+                'id' => parent::makeCallId($rawCall, $jRec['id']),
                 'active_journey_id' => $jRec['id'],
                 'line_private_code' => $jRec['line_private_code'],
             ]
         ));
-    }
-
-    /**
-     * Query the raw netex data for a day's journey data
-     *
-     * @param string $date
-     *
-     * @return \Illuminate\Support\Collection
-     */
-    protected function getRawJourneys($date)
-    {
-        return DB::table('netex_vehicle_journeys', 'journey')
-            ->select([
-                'journey.id as vehicle_journey_id',
-                'journey.name',
-                'journey.private_code',
-                'route.direction',
-                'journey.journey_pattern_ref',
-                'journey.operator_ref as operator_id',
-                'line.id as line_id',
-                'line.public_code as line_public_code',
-                'line.private_code as line_private_code',
-                'line.name as line_name',
-                'line.transport_mode',
-                'line.transport_submode',
-            ])
-            ->join('netex_journey_patterns as pattern', 'journey.journey_pattern_ref', '=', 'pattern.id')
-            ->join('netex_routes as route', 'pattern.route_ref', 'route.id')
-            ->join('netex_lines as line', 'journey.line_ref', 'line.id')
-            ->whereIn('journey.calendar_ref', function (\Illuminate\Database\Query\Builder $query) use ($date) {
-                $query->select('ref')->from('netex_calendar')->whereDate('date', '=', $date);
-            })
-            ->orderBy('line.private_code')
-            ->orderBy('journey.private_code')
-            ->get();
-    }
-
-    /**
-     * @param string $journeyRef
-     *
-     * @return \Illuminate\Support\Collection
-     */
-    protected function getRawCalls($journeyRef)
-    {
-        $ret = DB::table('netex_passing_times', 'ptime')
-            ->select([
-                'ptime.arrival_time',
-                'ptime.departure_time',
-                'patstop.alighting',
-                'patstop.boarding',
-                'patstop.order',
-                'ddisp.front_text as destination_display',
-                'stopass.quay_ref as stop_quay_id',
-                'quay.privateCode as quay_private_code',
-                'quay.publicCode as quay_public_code',
-                'stop.name as stop_place_name',
-            ])
-            ->join('netex_journey_pattern_stop_point as patstop', 'ptime.journey_pattern_stop_point_ref', '=', 'patstop.id')
-            ->leftJoin('netex_destination_displays as ddisp', 'patstop.destination_display_ref', '=', 'ddisp.id')
-            ->join('netex_stop_assignments as stopass', 'patstop.stop_point_ref', '=', 'stopass.id')
-            ->join('netex_stop_quay as quay', 'stopass.quay_ref', '=', 'quay.id')
-            ->join('netex_stop_place as stop', 'quay.stop_place_id', '=', 'stop.id')
-            ->where('ptime.vehicle_journey_ref', '=', $journeyRef)
-            ->orderBy('patstop.order')
-            ->get();
-        return $ret;
     }
 
     /**
